Separate logic of creating chat rooms, ChatWindow should create it and ExploreUsersBox component should not have anything to do with it
