+++ frontend/src/components/ExploreUsersBox.jsx
@@ -2,37 +2,19 @@ import { useContext } from "react";
 import { UserContext } from "./UserContext";
 import UserAvatar from "./UserAvatar";
 import { ArrowRightIcon } from "@chakra-ui/icons";
-import { Skeleton, SkeletonCircle, SkeletonText, Box } from "@chakra-ui/react";
+import { SkeletonCircle, SkeletonText, Box } from "@chakra-ui/react";
 
 export default function ExploreUsersBox({
   setClickedPage,
   setCurrentChatUser,
-  setChatRoom,
 }) {
   const { userData, loading } = useContext(UserContext);
 
   const handleChatClick = (clickedUser) => {
     setCurrentChatUser(clickedUser);
-    createChatRoom(clickedUser);
     setClickedPage("messages");
   };
 
-  const createChatRoom = async (user) => {
-    const url = "http://localhost:3000/api/v1/chat_rooms";
-    const currentUser = userData.current_user;
-    const body = {
-      chat_room: {
-        user_ids: [user.id, currentUser.id],
-      },
-    };
-    await fetch(url, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(body),
-    });
-  };
   return (
     <>
       {loading ? (
