+++ frontend/src/routes/Homepage/Homepage.jsx
@@ -11,13 +11,15 @@ export default function Homepage() {
   // State to change the page displayed to user
   const [clickedPage, setClickedPage] = useState("explore");
   const [currentChatUser, setCurrentChatUser] = useState(null);
-  const [chatRoom, setChatRoom] = useState(null);
   return (
     <UserProvider>
       <div className="py-5 flex justify-center items-center h-screen w-screen bg-gradient-to-b from-gradient-top to-gradient-bot">
         <div className="p-5 gap-x-5 gap-y-3 grid grid-cols-12 grid-rows-12 w-11/12 bg-gradient-to-b from-gradient-box-top to-gradient-box-bot self-center min-h-96 max-h-full bg-black rounded-xl">
           <div className="flex justify-center items-center rounded-xl row-start-11 col-start-1 col-span-12 row-span-3 md:row-start-11 md:row-span-2 md:col-start-1 md:col-span-12 lg:col-start-1 lg:col-span-1 lg:row-span-12 bg-messenger-blue">
-            <Sidebar setClickedPage={setClickedPage} />
+            <Sidebar
+              setClickedPage={setClickedPage}
+              currentChatUser={currentChatUser}
+            />
           </div>
           <div className="col-span-12 md:col-span-12 lg:block lg:col-start-2 lg:col-span-3 rounded-xl">
             <SearchBar />
@@ -31,18 +33,13 @@ export default function Homepage() {
                 <ExploreUsersBox
                   setCurrentChatUser={setCurrentChatUser}
                   setClickedPage={setClickedPage}
-                  setChatRoom={setChatRoom}
                 />
               )}
               {clickedPage === "profile" && (
                 <Profile setClickedPage={setClickedPage} />
               )}
               {clickedPage === "messages" && (
-                <ChatWindow
-                  setClickedPage={setClickedPage}
-                  chatRoom={chatRoom}
-                  currentChatUser={currentChatUser}
-                />
+                <ChatWindow currentChatUser={currentChatUser} />
               )}
             </>
           </div>
